# docs/openapi/spec/membershipfees.yaml
...
openapi: 3.1.0
info:
  title: Klabis API — Membership fees module
  description: The Membership fees slice of the Klabis API. See klabis.yaml for the full API.
  version: 0.1.0
servers:
  - url: https://localhost:8443
    description: Local development server
  - url: https://api.klabis.com
    description: Production server
paths:
  /api/membership-fee-tiers:
    get:
      operationId: listTiers
      tags: [MembershipFeeTiers]
      summary: List all membership fee tiers
      description: Returns all membership fee tiers in the catalog.
      security:
        - KlabisAuth: [MEMBERS]
      # No x-klabis-authority: readable by any authenticated caller. The activeCampaign/
      # pastCampaigns links and the createTier/publishYear affordances are only added for callers
      # with MEMBERS:MANAGE, but that is decided inline in the controller (isAdmin()) rather than
      # gating the whole operation.
      responses:
        '200':
          description: List of membership fee tiers retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipFeeTierSummaryResponseList'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelMembershipFeeTierSummaryResponse'
          x-hal-links:
            self:
              description: This collection
            activeCampaign:
              operation: getPublication
              description: Present only for callers with MEMBERS:MANAGE, when a campaign is active
            pastCampaigns:
              operation: listPublications
              description: Present only for callers with MEMBERS:MANAGE
          x-hal-templates:
            createTier:
              operation: createTier
              description: Present only for callers with MEMBERS:MANAGE
            publishYear:
              operation: publishYear
              description: Present only for callers with MEMBERS:MANAGE
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    post:
      operationId: createTier
      tags: [MembershipFeeTiers]
      summary: Create a membership fee tier
      description: Creates a new membership fee tier in the catalog.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      requestBody:
        required: true
        x-codegen-request-body-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateMembershipFeeTierRequest'
      responses:
        '201':
          description: Membership fee tier created
          headers:
            Location:
              description: URI of the created tier
              schema:
                type: string
                format: uri
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
  /api/membership-fee-tiers/{id}:
    get:
      operationId: getTier
      tags: [MembershipFeeTiers]
      summary: Get membership fee tier details
      description: Returns a single membership fee tier by ID.
      security:
        - KlabisAuth: [MEMBERS]
      # No x-klabis-authority: readable by any authenticated caller. editTier/deleteTier
      # affordances still require MEMBERS:MANAGE via klabisAfford's own authorization check.
      parameters:
        - $ref: '#/components/parameters/TierIdParam'
      responses:
        '200':
          description: Membership fee tier found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MembershipFeeTierResponse'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/EntityModelMembershipFeeTierResponse'
          x-hal-links:
            self:
              description: This tier
            collection:
              operation: listTiers
              description: Back to the tier list
            rules:
              operation: listRules
              description: Payment rules for this tier
          x-hal-templates:
            default:
              operation: editTier
              description: Present only for callers with MEMBERS:MANAGE
            deleteTier:
              operation: deleteTier
              description: Present only for callers with MEMBERS:MANAGE
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    patch:
      operationId: editTier
      tags: [MembershipFeeTiers]
      summary: Edit a membership fee tier
      description: |
        Partial update of a tier's name, yearly fee and/or payment rules. Fields omitted from the
        request are left unchanged.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      parameters:
        - $ref: '#/components/parameters/TierIdParam'
      requestBody:
        required: true
        x-codegen-request-body-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditMembershipFeeTierRequest'
      responses:
        '204':
          description: Membership fee tier updated
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    delete:
      operationId: deleteTier
      tags: [MembershipFeeTiers]
      summary: Delete a membership fee tier
      description: Deletes a membership fee tier from the catalog.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      parameters:
        - $ref: '#/components/parameters/TierIdParam'
      responses:
        '204':
          description: Membership fee tier deleted
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
  /api/membership-fee-tiers/{id}/rules:
    get:
      operationId: listRules
      tags: [MembershipFeeTiers]
      summary: List payment rules for a membership fee tier
      description: Returns all payment rules defined on a tier.
      security:
        - KlabisAuth: [MEMBERS]
      # No x-klabis-authority: readable by any authenticated caller. addRule affordance still
      # requires MEMBERS:MANAGE via klabisAfford's own authorization check.
      parameters:
        - $ref: '#/components/parameters/TierIdParam'
      responses:
        '200':
          description: List of payment rules retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRuleResponseList'
            application/prs.hal-forms+json:
              schema:
                $ref: '#/components/schemas/CollectionModelEntityModelPaymentRuleResponse'
          x-hal-links:
            self:
              description: This collection
          x-hal-templates:
            addRule:
              operation: addRule
              description: Present only for callers with MEMBERS:MANAGE
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'

    post:
      operationId: addRule
      tags: [MembershipFeeTiers]
      summary: Add a payment rule to a membership fee tier
      description: |
        Adds a new payment rule (percentage or fixed amount) for an event type/ranking combination.
        Returns Location header pointing to the created rule.
      security:
        - KlabisAuth: [MEMBERS]
      x-klabis-authority: MEMBERS_MANAGE
      parameters:
        - $ref: '#/components/parameters/TierIdParam'
      requestBody:
        required: true
        x-codegen-request-body-name: request
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddPaymentRuleRequest'
      responses:
        '201':
          description: Payment rule created
          headers:
            Location:
              description: URI of the created rule
              schema:
                type: string
                format: uri
          content:
            application/prs.hal-forms+json: {}
        '400':
          $ref: './_shared/responses.yaml#/components/responses/BadRequest'
        '401':
          $ref: './_shared/responses.yaml#/components/responses/Unauthorized'
        '403':
          $ref: './_shared/responses.yaml#/components/responses/Forbidden'
        '404':
          $ref: './_shared/responses.yaml#/components/responses/NotFound'
        '409':
          $ref: './_shared/responses.yaml#/components/responses/Conflict'
        '422':
          $ref: './_shared/responses.yaml#/components/responses/UnprocessableEntity'
...
components:
  securitySchemes:
    KlabisAuth:
      type: oauth2
      description: OAuth2 authentication with Klabis Authorization Server
      flows:
        authorizationCode:
          authorizationUrl: /oauth2/authorize
          tokenUrl: /oauth2/token
          scopes:
            openid: authenticate
            profile: display user profile information
            MEMBERS: Members
            EVENTS: Events
  parameters:
    TierIdParam:
      name: id
      in: path
      description: Membership fee tier UUID
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    MembershipFeeTierSummaryResponseList:
      type: array
      items:
        $ref: '#/components/schemas/MembershipFeeTierSummaryResponse'

    # Envelope schemas — never generated into Java; see klabis-api-spec skill:
    # "Payload and envelope are separate schemas".
    CollectionModelEntityModelMembershipFeeTierSummaryResponse:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            membershipFeeTierSummaryResponseList:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelMembershipFeeTierSummaryResponse'
        _links:
          $ref: './_shared/hal.yaml#/components/schemas/Links'
        _templates:
          $ref: './_shared/hal.yaml#/components/schemas/HalFormsTemplates'
...
    EntityModelMembershipFeeTierResponse:
      allOf:
        - $ref: '#/components/schemas/MembershipFeeTierResponse'
        - type: object
          properties:
            _links:
              $ref: './_shared/hal.yaml#/components/schemas/Links'
            _templates:
              $ref: './_shared/hal.yaml#/components/schemas/HalFormsTemplates'
    MembershipFeeTierResponse:
      type: object
      description: Complete membership fee tier information.
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        yearlyFeeAmount:
          type: number
        yearlyFeeCurrency:
          type: string
    CreateMembershipFeeTierRequest:
      type: object
      required: [name, yearlyFeeAmount]
      properties:
        name:
          type: string
          x-klabis-not-blank: true
        yearlyFeeAmount:
          type: number
          exclusiveMinimum: 0
        yearlyFeeCurrency:
          type: string
          description: Defaults to CZK when omitted.
    EditMembershipFeeTierRequest:
      type: object
      description: |
        Partial update. Fields omitted from the request are left unchanged. When rules is
        provided, it replaces the tier's entire rule set.
      properties:
        name:
          type: string
        yearlyFeeAmount:
          type: number
        yearlyFeeCurrency:
          type: string
        rules:
          type: array
          items:
            $ref: '#/components/schemas/PaymentRuleRequest'
...
    PaymentRuleResponseList:
      type: array
      items:
        $ref: '#/components/schemas/PaymentRuleResponse'
    CollectionModelEntityModelPaymentRuleResponse:
      type: object
      properties:
        _embedded:
          type: object
          properties:
            paymentRuleResponseList:
              type: array
              items:
                $ref: '#/components/schemas/EntityModelPaymentRuleResponse'
        _links:
          $ref: './_shared/hal.yaml#/components/schemas/Links'
        _templates:
          $ref: './_shared/hal.yaml#/components/schemas/HalFormsTemplates'
...
    AddPaymentRuleRequest:
      type: object
      required: [eventTypeId, rankingShortName, ruleType]
      properties:
        eventTypeId:
          type: string
          format: uuid
        rankingShortName:
          type: string
          x-klabis-not-blank: true
        ruleType:
          type: string
          x-klabis-not-blank: true
        percentage:
          type: integer
          format: int32
          description: Required when ruleType is PERCENTAGE.
        fixedAmount:
          type: number
          description: Required when ruleType is FIXED_AMOUNT.
        fixedCurrency:
          type: string
          description: Defaults to CZK when omitted.
...